+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - A new prop `page` has been added. It allows to wholly [control](https://reactjs.org/docs/forms.html#controlled-components) the component. (see https://github.com/AdeleD/react-paginate/issues/124)
 - The `forcePage` prop is now DEPRECATED. It was allowing a subpar controlled mode, which confused users ([#124](https://github.com/AdeleD/react-paginate/issues/124) [#374](https://github.com/AdeleD/react-paginate/issues/374) [#356](https://github.com/AdeleD/react-paginate/issues/356)) and potentially created race conditions. Please use `page` prop instead.
+- The `disableInitialCallback` prop now default to `true` when the `page` prop is used (in controlled mode)
 - Add a warning when providing both `initialPage` and `page` props. You should only provides `page` when the component is [controlled](https://reactjs.org/docs/forms.html#controlled-components). (see: https://github.com/AdeleD/react-paginate/pull/290)
 - Removed support for depecrated `extraAriaContext` (please use `ariaLabelBuilder` instead)